refactored the files structure due to prevent memory leak error
