+++ packages/e2e-tests/specs/admin-event-list/bulkActions.test.ts
