+++ packages/e2e-tests/specs/admin-event-list/linkForViewAllEvents.test.ts
@@ -2,7 +2,7 @@ import { saveVideo, PageVideoCapture } from 'playwright-video';
 
 import { createNewEvent } from '@e2eUtils/admin/events';
 import { EventsListSurfer, Goto, DefaultSettingsManager } from '@e2eUtils/admin';
-import { eventData } from '../../../../shared/data';
+import { eventData } from '../shared/data';
 import { ElementHandle } from 'packages/e2e-tests/types';
 import { activatePlugin, deactivatePlugin } from '@e2eUtils/admin/wp-plugins-page';
