+++ packages/e2e-tests/specs/admin-events-categories/editCategory.test.ts
