+++ packages/e2e-tests/specs/admin-events-default-settings/defaultMaximumTickets.test.ts
