+++ packages/e2e-tests/specs/admin-events-default-settings/defaultRegistrationStatus.test.ts
@@ -1,6 +1,6 @@
 import { saveVideo, PageVideoCapture } from 'playwright-video';
 import { Goto, DefaultSettingsManager } from '@e2eUtils/admin';
-import { defaultSettingsData } from '../../../shared/data';
+import { defaultSettingsData } from '../shared/data';
 import { activatePlugin, deactivatePlugin } from '@e2eUtils/admin/wp-plugins-page';
 
 const baristaPlugin = 'barista/ee-barista.php';
