+++ packages/e2e-tests/specs/admin-events-editor-tickets-list/card.test.ts
