+++ packages/e2e-tests/specs/admin-events-editor-tickets-list/price-inline-vs-tpc.test.ts
