+++ packages/e2e-tests/specs/admin-events-editor-tickets-list/price-precision.test.ts
