+++ packages/e2e-tests/specs/admin-events-templates/archives/displayStatusBanner.test.ts
