+++ packages/e2e-tests/specs/admin-events-templates/archives/eventSlug.test.ts
