+++ packages/e2e-tests/specs/admin-events-templates/archives/resetEventListSettings.test.ts
