+++ packages/e2e-tests/specs/admin-events-templates/single/displayStatusBanner.test.ts
