+++ packages/e2e-tests/specs/admin-events-templates/single/displayVenueDetails.test.ts
