+++ packages/e2e-tests/specs/admin-events-templates/ticket-selector/dateAndTimeFilterMaxChecked.test.ts
