+++ packages/e2e-tests/specs/admin-events-templates/ticket-selector/moreTicketSelectorTenMaxAttendees.test.ts
