+++ packages/e2e-tests/specs/admin-events-templates/ticket-selector/showDateAndTimeFilter.test.ts
@@ -10,7 +10,7 @@ import {
 	EDTRGlider,
 	DefaultSettingsManager,
 } from '@e2eUtils/admin';
-import { eventData } from '../../../../shared/data';
+import { eventData } from '../../shared/data';
 import { activatePlugin, deactivatePlugin } from '@e2eUtils/admin/wp-plugins-page';
 
 const baristaPlugin = 'barista/ee-barista.php';
