+++ packages/e2e-tests/specs/admin-events-templates/ticket-selector/showTicketSaleInfo.test.ts
